backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
geoalchemy2==0.15.2
shapely==2.0.6
geopandas==1.0.1
pandas==2.2.3
networkx==3.3
cachetools==5.5.0
structlog==24.4.0
httpx==0.27.2
pyyaml==6.0.2
requests==2.32.3
reportlab==4.2.5
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
